put Person into seperate module
