+++ package.json
@@ -21,6 +21,7 @@
   "devDependencies": {
     "browserify": "^14.4.0",
     "eslint": "^4.1.1",
+    "eslint-plugin-node": "^5.1.0",
     "watch": "^1.0.2"
   }
 }
